+++ firmware/controllers/limp_manager.h
@@ -88,7 +88,7 @@ public:
 	ShutdownController shutdownController;
 
 	// This is called from periodicFastCallback to update internal state
-	void updateState(int rpm, efitick_t nowNt);
+	void updateState(float rpm, efitick_t nowNt);
 
 	void onFastCallback() override;
 	void onIgnitionStateChanged(bool ignitionOn) override;
