+++ unit_tests/tests/test_idle_controller.cpp
@@ -388,9 +388,9 @@ TEST(idle_v2, closedLoopDeadzone) {
 
 struct IntegrationIdleMock : public IdleController {
 	MOCK_METHOD(int, getTargetRpm, (float clt), (override));
-	MOCK_METHOD(ICP, determinePhase, (int rpm, int targetRpm, SensorResult tps, float vss, float crankingTaperFraction), (override));
+	MOCK_METHOD(ICP, determinePhase, (float rpm, float targetRpm, SensorResult tps, float vss, float crankingTaperFraction), (override));
 	MOCK_METHOD(float, getOpenLoop, (ICP phase, float rpm, float clt, SensorResult tps, float crankingTaperFraction), (override));
-	MOCK_METHOD(float, getClosedLoop, (ICP phase, float tps, int rpm, int target), (override));
+	MOCK_METHOD(float, getClosedLoop, (ICP phase, float tps, float rpm, float target), (override));
 	MOCK_METHOD(float, getCrankingTaperFraction, (), (const, override));
 };
